removed closest function test
